Split DaemonState in two
Put systemd stuff in SystemDState
